linked welcome popup to pg import
