+++ src/components/popup/exportPopUp.jsx
@@ -95,24 +95,9 @@ function ExportPopUp(props) {
               ipc.send('show-export-dialog', gqlSchema, gqlResolvers, sqlScripts, buildENV(uri), queries);
               dispatch({ type: SET_POP_UP, payload: '' });
             } else {
-
-              //TEST --------------------------------
-              //test to see if table import will log scraped pg data to console
-              async function importTables() {
-                const tables = {};
-                ipc.on('tables-imported', (event, arg) => {
-                  console.log("import tables, no async: ", arg)
-                  dispatch({ type: IMPORT_TABLES, payload: arg})
-                })
-                ipc.send('import-tables', tables, buildENV(uri));
-                // dispatch({ type: IMPORT_TABLES, payload: update });
-                // let update = await ipc.on('tables-imported', (event, arg) => arg);
-                // console.log("tables (render): ", update)
-                document.querySelector('#error').classList.remove('invisible')
-              }
-              importTables();
-              
-          }}}>
+              document.querySelector('#error').classList.remove('invisible')
+            }
+          }}>
           <Input label="Enter Postgres Connection URI" margin="normal" type="text" required />
           <p className="invisible" id="error">URI needs to start with "postgres://" or "postgresql://"</p>
           <DialogActionsDiv>
